CSS changes and style fixes
The admin dashboard was confusing so now each section is inside a
separate card. Also buttons in the appointments table match the colors
of status (Ex: green "Confirm" button for a green "Confirmed"
appointment). The changes are:

- Add cards to separate sections in admin dashboard
- Add style to buttons in appointments table

Also, the PR has style changes and refactorings to make the code more
clear and clean.

- Simplify partial render in views to short form.
- Add null appointment spec to avoid type checking in views.
- Move appointment methods from controller in appointment create action
  to appointment model to avoid fat controller.