Add admin ability to assign appointments
Admins can now assign appointments to customers. This feature completes
the admin creation of appointment without customers.
To assign an uncompleted appointment to a customer an admin can either
search for existing customers or create a new customer.
The search of customers in the appointment assignment page is done via
ajax request to keep the page with the current appointment on screen and
to be able to use that information in the appointment assignation.