Add calendar of appointments for admins
This feature gives admins the abiity to see a calendar with the active
appointments (booked and confirmed). Appointments appear in a badge with
the color corresponding to their status. This allows admins to see a
view of the work for the next weeks and assess the availability of the
shop when a customer tries to book an appointment.

The commit contains the following changes:

- Add simple_calendar gem to generate the calendar views.
- Generate different calendar views in "simple_calendar" folder to
  customize them.
- Customize monthly calendar view to show the right first day of the
  week and table header based on Argentina's culture.
- Localize and add translation for days and months name in Spanish.
- Add calendar resource