Fix bug on appointments creation
Before create callbacks were not working on appointments creation
leaving the duration and status field with incorrect values. This
generated an error of missing fields when listing those created
appointments.
Adding "self" to the methods called before creating the record solves
the problem.

This commit also adds Procfile changes to heroku deployment and
stylesheets changes for badges shown in the calendar appointments.