Rename appointment_assignments to _completions
There is already an assignments controller which which models the
concept of a work scheduled to do in a vehicle. On the other hand,
appointment_assignments were used to complete uncompleted appointments
by assigning them a vehicle (when admins create the appointment before
selecting o creatin a customer/vehicle). As the work "assignment" was
being used for two different concepts and the latter one is also
completing the appointment, all the related elements in the app were
renamed to appointment_completions.