Add link to edit an assignment in appointment details
An admin can access to edit an assignment from the appointment details
view. Before this commit there was no way to access the edit assignment
page except for going to the route by writing the right assignment id in
the url.
To access the appointment details this commit adds a link to each one of
the appointments in appointments table.